+++ crates/wdl-format/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* `quote_style` configuration option, with choices: `double`, `single`, `preserve` ([#1200](https://github.com/stjude-rust-labs/sprocket/pull/1200))
 * Short literal arrays, if/then/else clauses, and symbolic import clauses are kept on the same line instead of being unconditionally line split ([#1154](https://github.com/stjude-rust-labs/sprocket/pull/1154)).
 
 #### Fixed